+++ test/fakeSchemaDb2.js
@@ -1,5 +1,7 @@
+const { schema2 } = require('./helpers/schema2');
+
 exports.default = ({ types: t }) => ({
-	table: 'yass_test2.yass_test3',
+	table: `${schema2()}.yass_test3`,
 	schema: {
 		id: t.uuidKey,
 		name: t.string,